Fix validation errors and equipment not showing in dropdown
Dropdown was not showing value because data from server is set by reference, not copied, so when data variable is out of scope, certain fields are undefined, including equipmentId. Validation errors were also showing up because of required fields in non-CCE section. Instead of hiding the fields, they are removed (from ng-show to ng-if).